[Ingestion Client] Use Text Analytics Nuget + Async API (#1258)
* add text analytics nuget

* retry on 429

* remove min precision

* update nugets

* update function to be long running

* Update TextAnalyticsProvider.cs

* Update TextAnalyticsProvider.cs

+++ samples/ingestion/ingestion-client/Connector/Serializable/TranscriptionResult/Sentiment.cs
@@ -10,12 +10,12 @@ namespace Connector
     public class Sentiment
     {
         [JsonProperty("negative")]
-        public float Negative { get; set; }
+        public double Negative { get; set; }
 
         [JsonProperty("neutral")]
-        public float Neutral { get; set; }
+        public double Neutral { get; set; }
 
         [JsonProperty("positive")]
-        public float Positive { get; set; }
+        public double Positive { get; set; }
     }
 }
